삭제시 확인창 구현

+++ src/main/webapp/jsp/article/list.jsp
@@ -36,7 +36,7 @@ int totalPage = (int)request.getAttribute("totalPage");
 			<td><%= (int)articleRow.get("id") %></td>
 			<td><%= (LocalDateTime)articleRow.get("regDate") %></td>
 			<td><a href="detail?id=<%= (int)articleRow.get("id") %>"><%= (String)articleRow.get("title") %></a></td>
-			<td><a href="doDelete?id=<%= (int)articleRow.get("id") %>">삭제버튼</a></td>
+			<td><a href="doDelete?id=<%= (int)articleRow.get("id") %>" onclick="if(confirm('정말 삭제하시겠습니까?')==false) return false;">삭제버튼</a></td>
 		</tr>
 	<% } %>
 	</table>
@@ -48,6 +48,12 @@ int totalPage = (int)request.getAttribute("totalPage");
 	</style>
 	
 	<div class="page">
+	
+	<!--  -->
+	<%if(totalPage>11){ %>
+	
+	<%} %>
+	
 		<%for(int i = 1; i <= totalPage; i++){ %>
 			<a class="<%= cPage == i ? "red" : "" %>" href="list?page=<%= i %>"><%= i %></a>
 		<%} %>
